refactor : AttackEntity 빌더 패턴 없애고, 팩토리 함수 추가
